feat(server): admin monitoring/moderation, task lifecycle, templates, observability
- Admin: /admin/stats|activity|users + requireAdmin + user moderation
  (suspend/role, session-kill); make-admin script; created the missing
  activity_logs table so the audit trail finally records
- Task lifecycle: edit/cancel, two-party completion handshake
  (submit / request-changes / confirm), data-driven categories,
  recurring + templated tasks (templates CRUD + sendRecurring job)
- Observability: structured access logging, captureException + crash handlers,
  DB index audit (tasks(assigned_to), feed composite, pg_trgm search indexes)
- Fixes: businesses POST 500 (fee_paid param type), activity_logs insert
  type error, login blocks suspended/deleted accounts
- Migrations 13-20; ~40 new tests (125 total)

+++ server/routes/businesses.js
@@ -163,8 +163,8 @@ router.post('/',
            (name, category, description, address, map_hint, phone, whatsapp, email,
             hours, image_urls, logo_url, link_url, status, fee_paid, paid_at,
             signed_by_rep, notes)
-         VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14,
-                 CASE WHEN $14 IS NOT NULL THEN NOW() ELSE NULL END,$15,$16)
+         VALUES ($1,$2,$3,$4,$5,$6,$7,$8,$9,$10,$11,$12,$13,$14::numeric,
+                 CASE WHEN $14::numeric IS NOT NULL THEN NOW() ELSE NULL END,$15,$16)
          RETURNING *`,
         [b.name, b.category, b.description || null, b.address || null, b.mapHint || null,
          b.phone || null, b.whatsapp || null, b.email || null, b.hours || null,
